demo/openai-agents-sdk-demo/pyproject.toml
[project]
name = "halo-openai-agents-demo"
version = "0.1.0"
description = "HALO demo: OpenAI Agents SDK writing inference.net-format JSONL traces"
requires-python = ">=3.12"
dependencies = [
    "openai-agents",
    "typer",
    "rich",
    "python-dotenv",
]
